ensure extents updated by SyncToDisk(), and that new features are merged into existing extents (https://trac.osgeo.org/gdal/ticket/4079)
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@22346 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/oci/ogrocisession.cpp
@@ -384,7 +384,7 @@ OGROCISession::GetParmInfo( OCIParam *hParmDesc, OGRFieldDefn *poOGRDefn,
         CPLError( CE_Failure, CPLE_AppDefined, 
                   "Column length (%d) longer than column name buffer (%d) in\n"
                   "OGROCISession::GetParmInfo()", 
-                  nColLen, sizeof(szTermColName) );
+                  nColLen, (int) sizeof(szTermColName) );
         return CE_Failure;
     }
